+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Improved UX when data loading from a file by showing progress bar
 - Added `App#loadDataFromStream()` and `App#loadDataFromFile()` methods
 - Added `unsubscribe` as a second parameter for publisher's subscriber handler
+- Changed `source` view to highligh using CodeMirror only
+- Enabled `source` view syntaxes can be observed on `#views-showcase:source` page
+- Removed `Prismjs` and related dependencies
 
 ## 1.0.0-beta.52 (19-11-2020)